+++ frontend/src/scenes/insights/__stories__/trendsTable.json
@@ -1,0 +1,103 @@
+{
+    "id": 51,
+    "short_id": "jEAIVJnV",
+    "name": "",
+    "derived_name": "Pageview count",
+    "filters": {
+        "events": [
+            {
+                "id": "$pageview",
+                "name": "$pageview",
+                "type": "events",
+                "order": 0
+            }
+        ],
+        "actions": [],
+        "display": "ActionsTable",
+        "insight": "TRENDS",
+        "interval": "day",
+        "properties": {
+            "type": "AND",
+            "values": []
+        },
+        "filter_test_accounts": false
+    },
+    "filters_hash": "cache_7b48d5b511fdab675b9cf64c4c8aef09",
+    "order": null,
+    "deleted": false,
+    "dashboard": null,
+    "layouts": {},
+    "color": null,
+    "last_refresh": null,
+    "refreshing": false,
+    "result": [
+        {
+            "action": {
+                "id": "$pageview",
+                "type": "events",
+                "order": 0,
+                "name": "$pageview",
+                "custom_name": null,
+                "math": null,
+                "math_property": null,
+                "math_group_type_index": null,
+                "properties": {}
+            },
+            "label": "$pageview",
+            "count": 0,
+            "data": [],
+            "labels": [],
+            "days": [
+                "2022-03-08",
+                "2022-03-09",
+                "2022-03-10",
+                "2022-03-11",
+                "2022-03-12",
+                "2022-03-13",
+                "2022-03-14",
+                "2022-03-15"
+            ],
+            "aggregated_value": 47258,
+            "filter": {
+                "date_from": "2022-03-08T00:00:00+00:00",
+                "date_to": "2022-03-15T21:38:06.869054+00:00",
+                "display": "ActionsTable",
+                "events": "[{\"id\": \"$pageview\", \"type\": \"events\", \"order\": 0, \"name\": \"$pageview\", \"custom_name\": null, \"math\": null, \"math_property\": null, \"math_group_type_index\": null, \"properties\": {}}]",
+                "insight": "TRENDS",
+                "interval": "day"
+            },
+            "persons": {
+                "filter": {
+                    "entity_id": "$pageview",
+                    "entity_type": "events",
+                    "entity_math": null
+                },
+                "url": "api/projects/1/actions/people/?date_from=2022-03-08T00%3A00%3A00%2B00%3A00&date_to=2022-03-15T21%3A38%3A06.869054%2B00%3A00&display=ActionsTable&events=%5B%7B%22id%22%3A+%22%24pageview%22%2C+%22type%22%3A+%22events%22%2C+%22order%22%3A+0%2C+%22name%22%3A+%22%24pageview%22%2C+%22custom_name%22%3A+null%2C+%22math%22%3A+null%2C+%22math_property%22%3A+null%2C+%22math_group_type_index%22%3A+null%2C+%22properties%22%3A+%7B%7D%7D%5D&insight=TRENDS&interval=day&entity_id=%24pageview&entity_type=events"
+            }
+        }
+    ],
+    "created_at": "2022-03-15T21:31:00.192917Z",
+    "created_by": {
+        "id": 1,
+        "uuid": "017f7913-c1ee-0000-faff-0fe5ace3849a",
+        "distinct_id": "Q1OD2NmnqqbqrLR45mYm2p1VNNbSO9DBFsMig90GWkK",
+        "first_name": "Marius",
+        "email": "marius@posthog.com"
+    },
+    "description": "",
+    "updated_at": "2022-03-15T21:38:08.043665Z",
+    "tags": [],
+    "favorited": false,
+    "saved": true,
+    "last_modified_at": "2022-03-15T21:38:08.039718Z",
+    "last_modified_by": {
+        "id": 1,
+        "uuid": "017f7913-c1ee-0000-faff-0fe5ace3849a",
+        "distinct_id": "Q1OD2NmnqqbqrLR45mYm2p1VNNbSO9DBFsMig90GWkK",
+        "first_name": "Marius",
+        "email": "marius@posthog.com"
+    },
+    "is_sample": false,
+    "effective_restriction_level": 21,
+    "effective_privilege_level": 37
+}
